src/backend/backend_src/requirements.txt
absl-py==1.2.0
altgraph==0.17.2
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
convertdate==2.4.0
cycler==0.11.0
finnhub-python==2.4.13
Flask==2.1.3
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.0
h5py==3.7.0
hijri-converter==2.2.4
holidays==0.14.2
idna==3.3
importlib-metadata==4.12.0
influxdb-client==1.31.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
libclang==14.0.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
multitasking==0.0.11
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyMeeus==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==3.2.0
scikit-learn==1.1.1
scipy==1.9.0
six==1.16.0
sklearn==0.0
src==0.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing-extensions==4.3.0
urllib3==1.26.11
websocket-client==1.3.3
Werkzeug==2.2.1
wrapt==1.14.1
yfinance==0.1.74
zipp==3.8.1
